refactor(m5/migration): extract shared SQL-AST helpers into sql_ast module
Move seven `pub(crate)` AST helpers from `migration/tier1.rs` into a new
sibling module `migration/sql_ast.rs`. The helpers manipulate
`sqlparser::ast::{Expr, Select, Statement, TableWithJoins}` and are called
from both `tier1.rs` and `tier2.rs` at near-even rates (37 / 33 references
per #587's grep) — tier2 was implicitly importing them from tier1, which is
the wrong direction for sibling modules.

Helpers moved verbatim:
  - from_is_events_with_optional_exposures_join
  - from_is_exposures_with_metric_events_join
  - find_join_on_expr
  - extract_select
  - collect_and_predicates
  - strip_table_alias
  - unwrap_nested

The private `table_name_from_factor_str` follows since it's the only
caller of `TableFactor` in this set and is used exclusively by the two
`from_is_*` helpers. Three now-unused `sqlparser::ast` imports
(`JoinConstraint`, `JoinOperator`, `SetExpr`) are dropped from tier1.rs.

Pure code-motion refactor — no behavior change. All 335 unit + integration
tests pass; `cargo clippy --no-deps` is clean. `grep -rn
"use.*migration::tier1::" crates/` now returns empty (tier2 imports from
sql_ast instead), matching the acceptance criterion in the issue.

Closes #587
